refactor(designer): Moved persistense directory into infrastructure
