+++ module/designer/src/Infrastructure/Persistence/Projector/DbalTemplateElementAddedEventProjector.php
@@ -7,15 +7,15 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\Designer\Persistence\Dbal\Projector;
+namespace Ergonode\Designer\Infrastructure\Persistence\Projector;
 
 use Doctrine\DBAL\Connection;
-use Ergonode\Designer\Domain\Event\TemplateElementChangedEvent;
+use Ergonode\Designer\Domain\Event\TemplateElementAddedEvent;
 use JMS\Serializer\SerializerInterface;
 
 /**
  */
-class TemplateElementChangedEventProjector
+class DbalTemplateElementAddedEventProjector
 {
     private const ELEMENT_TABLE = 'designer.template_element';
 
@@ -42,20 +42,18 @@ class TemplateElementChangedEventProjector
     /**
      * {@inheritDoc}
      */
-    public function __invoke(TemplateElementChangedEvent $event): void
+    public function __invoke(TemplateElementAddedEvent $event): void
     {
         $element = $event->getElement();
-        $this->connection->update(
+        $this->connection->insert(
             self::ELEMENT_TABLE,
-            [
-                'width' => $element->getSize()->getWidth(),
-                'height' => $element->getSize()->getHeight(),
-                'properties' => $this->serializer->serialize($element->getProperties(), 'json'),
-            ],
             [
                 'template_id' => $event->getAggregateId()->getValue(),
                 'x' => $element->getPosition()->getX(),
                 'y' => $element->getPosition()->getY(),
+                'width' => $element->getSize()->getWidth(),
+                'height' => $element->getSize()->getHeight(),
+                'properties' => $this->serializer->serialize($element->getProperties(), 'json'),
             ]
         );
     }
