+++ module/designer/src/Infrastructure/Persistence/Projector/DbalTemplateGroupChangedEventProjector.php
@@ -7,14 +7,14 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\Designer\Persistence\Dbal\Projector;
+namespace Ergonode\Designer\Infrastructure\Persistence\Projector;
 
 use Doctrine\DBAL\Connection;
-use Ergonode\Designer\Domain\Event\TemplateImageChangedEvent;
+use Ergonode\Designer\Domain\Event\TemplateGroupChangedEvent;
 
 /**
  */
-class TemplateImageChangedEventProjector
+class DbalTemplateGroupChangedEventProjector
 {
     private const TABLE = 'designer.template';
 
@@ -34,12 +34,12 @@ class TemplateImageChangedEventProjector
     /**
      * {@inheritDoc}
      */
-    public function __invoke(TemplateImageChangedEvent $event): void
+    public function __invoke(TemplateGroupChangedEvent $event): void
     {
         $this->connection->update(
             self::TABLE,
             [
-                'image_id' => $event->getTo()->getValue(),
+                'template_group_id' => $event->getNew()->getValue(),
             ],
             [
                 'id' => $event->getAggregateId()->getValue(),
