+++ module/designer/src/Infrastructure/Persistence/Projector/DbalTemplateImageAddedEventProjector.php
@@ -7,14 +7,14 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\Designer\Persistence\Dbal\Projector;
+namespace Ergonode\Designer\Infrastructure\Persistence\Projector;
 
 use Doctrine\DBAL\Connection;
-use Ergonode\Designer\Domain\Event\TemplateDefaultLabelChangedEvent;
+use Ergonode\Designer\Domain\Event\TemplateImageAddedEvent;
 
 /**
  */
-class TemplateDefaultTextChangedEventProjector
+class DbalTemplateImageAddedEventProjector
 {
     private const TABLE = 'designer.template';
 
@@ -34,12 +34,12 @@ class TemplateDefaultTextChangedEventProjector
     /**
      * {@inheritDoc}
      */
-    public function __invoke(TemplateDefaultLabelChangedEvent $event): void
+    public function __invoke(TemplateImageAddedEvent $event): void
     {
         $this->connection->update(
             self::TABLE,
             [
-                'default_label' => $event->getTo()->getValue(),
+                'image_id' => $event->getImageId()->getValue(),
             ],
             [
                 'id' => $event->getAggregateId()->getValue(),
