+++ module/designer/src/Infrastructure/Persistence/Repository/Decorator/CacheTemplateRepositoryDecorator.php
@@ -7,7 +7,7 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\Designer\Persistence\Dbal\Repository\Decorator;
+namespace Ergonode\Designer\Infrastructure\Persistence\Repository\Decorator;
 
 use Ergonode\Designer\Domain\Entity\Template;
 use Ergonode\Designer\Domain\Repository\TemplateRepositoryInterface;
